Include Android logger as static lib
Include the Android logger as a static lib to ensure it's present on
device when a binary tries to use it.

This is a CP of https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/99120 to reduce the
sync diff.

Bug: 221174084
Test: Run test locally
Merged-In: I4d0d98f91bbcc7bb4b9f1d85e05524e72f0d8d9d
Change-Id: I4d0d98f91bbcc7bb4b9f1d85e05524e72f0d8d9d