Remove sub-directory OWNERS from external/pigweed
Having Pigweed OWNERS in external/pigweed fails the
code owners check when uploading a code sync.
Since no code development occurs in this repo as
this is a mirror of the official Pigweed repo,
the main OWNERS file is enough to enable periodic
syncs.

This CL also removes arthuri and adds more OWNERS.

Bug: 277108894
Test: Presubmits
Change-Id: If460187a67018b334dbac5dc5e3a9a9c386ff6bd